implement some more PA API; v5: pamixer needs fixing, firefox-52 works
PAmixer eats 100% of CPU because it fails to check for errors properly.

Firefox works, but it's pretty accidental, in Firefox 52 they, too, fail to
properly check for errors ("PA is better simpler API" LOL). It doesn't loop with
100% CPU utilization like PAmixer by pure accident: they check context status
while in the loop instead with callbacks.

Even more interestingly, Firefox 58 has exactly the same bug, but it's not
triggered. Probably because in 58 that code became dead as in 58 they introduced
a lot of new audio handling code tied to Rust (and that code does check for
errors).

Still, "we can't support native ALSA, too tired of fixing bugs there" and then
rewriting all of the PA code is kinda hypocritical.